Move bidiagonal decomposition into its own module
